requirements.txt
pdfplumber
tiktoken
langchain
langchain-community
langchain-openai
langchain-groq
python-dotenv
langchain-huggingface
ragas
datasets
langchain-pinecone
streamlit
